+++ ios/Classes/MultiQrTrackerPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface MultiQrTrackerPlugin : NSObject<FlutterPlugin>
+@end